+++ amethyst_renderer/src/formats/mesh.rs
@@ -20,6 +20,9 @@ pub enum MeshData {
     /// Position and color
     PosColor(Vec<PosColor>),
 
+    /// Position, color and normal
+    PosColorNorm(Vec<PosColorNorm>),
+
     /// Position and texture coordinates
     PosTex(Vec<PosTex>),
 
@@ -41,6 +44,12 @@ impl Component for MeshData {
 impl From<Vec<PosColor>> for MeshData {
     fn from(data: Vec<PosColor>) -> Self {
         MeshData::PosColor(data)
+    }
+}
+
+impl From<Vec<PosColorNorm>> for MeshData {
+    fn from(data: Vec<PosColorNorm>) -> Self {
+        MeshData::PosColorNorm(data)
     }
 }
 
@@ -187,6 +196,10 @@ fn from_data(obj_set: ObjSet) -> Vec<PosNormTex> {
 pub fn create_mesh_asset(data: MeshData, renderer: &mut Renderer) -> Result<ProcessingState<Mesh>> {
     let data = match data {
         MeshData::PosColor(ref vertices) => {
+            let mb = MeshBuilder::new(vertices);
+            renderer.create_mesh(mb)
+        }
+        MeshData::PosColorNorm(ref vertices) => {
             let mb = MeshBuilder::new(vertices);
             renderer.create_mesh(mb)
         }