+++ docs/CHANGELOG.md
@@ -9,7 +9,8 @@ The format is based on [Keep a Changelog][kc], and this project adheres to
 
 ## Unreleased
 ### Added
-* Added JsonFormat ([#950]).
+* Support for debug lines using `DebugLines` pass, and `DebugLines` component or resource. ([#917], [#957])
+* Added JsonFormat ([#950]). 
 * `SpriteRender` pass to draw sprites without using `Material` and `Mesh`. ([#829], [#830])
 * Sprite animation uses the `SpriteRenderChannel`. ([#829], [#830])
 * State::handle_event can now handle multiple types of events. ([#887])
@@ -76,14 +77,17 @@ The format is based on [Keep a Changelog][kc], and this project adheres to
 [#904]: https://github.com/amethyst/amethyst/pull/904
 [#915]: https://github.com/amethyst/amethyst/pull/915
 [#868]: https://github.com/amethyst/amethyst/pull/868
+[#917]: https://github.com/amethyst/amethyst/issues/917
 [#933]: https://github.com/amethyst/amethyst/pull/933
 [#929]: https://github.com/amethyst/amethyst/pull/929
 [#934]: https://github.com/amethyst/amethyst/pull/934
 [#940]: https://github.com/amethyst/amethyst/pull/940
 [#946]: https://github.com/amethyst/amethyst/pull/946
 [#950]: https://github.com/amethyst/amethyst/pull/950
+[#957]: https://github.com/amethyst/amethyst/pull/957
 [winit_017]: https://github.com/tomaka/winit/blob/master/CHANGELOG.md#version-0172-2018-08-19
 [glutin_018]: https://github.com/tomaka/glutin/blob/master/CHANGELOG.md#version-0180-2018-08-03
+
 
 ## [0.8.0] - 2018-08
 ### Added